Improve touchpad/trackpad input precision
TOUCH_POINT no longer truncates input, input is rounded instead of floored just before sending mouse event, and DualSense data from SDL is scaled correctly. This improves "touchsticks" as well.

+++ JoyShockMapper/include/JoyShock.h
@@ -48,9 +48,9 @@ public:
 	// These two large functions are defined further down
 	void processStick(float stickX, float stickY, Stick &stick, float mouseCalibrationFactor, float deltaTime, bool &anyStickInput, bool &lockMouse, float &camSpeedX, float &camSpeedY);
 
-	void handleTouchStickChange(TouchStick &ts, bool down, short movX, short movY, float delta_time);
-	void handleLeftTouchStickChange(LeftTouchStick &ts, bool down, short movX, short movY, float delta_time);
-	void handleRightTouchStickChange(RightTouchStick &ts, bool down, short movX, short movY, float delta_time);
+	void handleTouchStickChange(TouchStick &ts, bool down, float movX, float movY, float delta_time);
+	void handleLeftTouchStickChange(LeftTouchStick &ts, bool down, float movX, float movY, float delta_time);
+	void handleRightTouchStickChange(RightTouchStick &ts, bool down, float movX, float movY, float delta_time);
 
 	bool hasVirtualController();
